refactor(daemon): extract assistant-message persistence from server.ts
Strangler-fig slice 3: move the assistant-message run-lifecycle persistence,
skill-plugin-candidate detection, and agent-event mapping helpers out of
server.ts into a new sibling module, byte-identical (sliced from source to
preserve raw Unicode: the × in the tool-loop message, em-dash comments, and the
SQL strings exactly).

Three sub-groups:
- Assistant-message lifecycle: pinAssistantMessageOnRunCreate,
  reconcileAssistantMessageOnRunEnd.
- Skill-plugin-candidate detection: isPluginAuthoringRun,
  hasGeneratedPluginArtifacts, assistantMessageEmittedQuestionForm,
  deferredSkillPluginCandidateForRun, detectSkillPluginCandidateOnRunSuccess,
  upsertSkillPluginCandidateAssistantMessage.
- Agent-event mapping: persistRunEventToAssistantMessage,
  runSseEventToPersistedAgentEvent, daemonAgentPayloadToPersistedAgentEvent,
  normalizePersistedToolInput.

All take db/runs/run as args; deps are imports (plugins/index, question-form-detect,
db, projects, crypto) -- no daemon module state, no cycle. server.ts imports back
the six it references and re-exports the three public ones.

No behavior change. Orphan/duplicate checks clean.
